+++ internal/cli/terminal_other.go
@@ -1,0 +1,7 @@
+//go:build !windows
+
+package cli
+
+import "os"
+
+func enableHelpVT(*os.File) func() { return func() {} }